Refactor instructor slot calendar
